DisentangledRep4RL/cluster_requirements.txt
numpy==1.22.4
absl-py==0.13.0
jupyterlab==3.0.14
scikit-image==0.18.1
termcolor==2.1.1
tb-nightly==2.11.0a20221013
imageio==2.19.3
imageio-ffmpeg==0.4.4
hydra-core==1.1.0
hydra-submitit-launcher==1.1.5
pandas==1.5.2
ipdb==0.13.9
yapf==0.31.0
mujoco_py==2.0.2.13
sklearn==0.0
matplotlib==3.5.1
opencv-python==4.6.0.66
wandb==0.13.5
pylint==2.15.9
distracting_control==0.1.1rc3

git+https://github.com/deepmind/dm_control.git
git+https://github.com/Farama-Foundation/d4rl@master#egg=d4rl

torch==1.9.0+cu111
torchvision==0.10.0+cu111
torchaudio==0.9.0